+++ src/promptflow/promptflow/_sdk/operations/_local_storage_operations.py
@@ -2,6 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 
+import copy
 import datetime
 import json
 import logging
@@ -39,7 +40,6 @@ from promptflow.contracts.run_info import RunInfo as NodeRunInfo
 from promptflow.contracts.run_info import Status
 from promptflow.contracts.run_mode import RunMode
 from promptflow.exceptions import UserErrorException
-from promptflow.executor._result import LineResult
 from promptflow.executor.flow_executor import BulkResult
 from promptflow.storage import AbstractRunStorage
 
@@ -137,13 +137,13 @@ class NodeRunRecord:
             lock.acquire()
             try:
                 with open(path, mode="w", encoding=DEFAULT_ENCODING) as f:
-                    json.dump(asdict(self), f)
+                    json.dump(asdict(self), f, ensure_ascii=False)
             finally:
                 lock.release()
         else:
             # for normal nodes in other line runs, directly write
             with open(path, mode="w", encoding=DEFAULT_ENCODING) as f:
-                json.dump(asdict(self), f)
+                json.dump(asdict(self), f, ensure_ascii=False)
 
 
 @dataclass
@@ -172,7 +172,7 @@ class LineRunRecord:
 
     def dump(self, path: Path) -> None:
         with open(path, mode="w", encoding=DEFAULT_ENCODING) as f:
-            json.dump(asdict(self), f)
+            json.dump(asdict(self), f, ensure_ascii=False)
 
 
 class LocalStorageOperations(AbstractRunStorage):
@@ -193,7 +193,10 @@ class LocalStorageOperations(AbstractRunStorage):
         self._flow_tools_json_path = (
             self._snapshot_folder_path / PROMPT_FLOW_DIR_NAME / LocalStorageFilenames.FLOW_TOOLS_JSON
         )
-        self._inputs_path = self._snapshot_folder_path / LocalStorageFilenames.INPUTS
+        self._inputs_path = self.path / LocalStorageFilenames.INPUTS  # keep this for other usages
+        # below inputs and outputs are dumped by SDK
+        self._sdk_inputs_path = self._inputs_path
+        self._sdk_output_path = self.path / LocalStorageFilenames.OUTPUTS
         # metrics
         self._metrics_path = self.path / LocalStorageFilenames.METRICS
         # legacy files: detail.json and outputs.jsonl(not the one in flow_outputs folder)
@@ -203,7 +206,7 @@ class LocalStorageOperations(AbstractRunStorage):
         # for normal node run records, store per node per line;
         # for reduce node run records, store centralized in 000000000.jsonl per node
         self.outputs_folder = self._prepare_folder(self.path / "flow_outputs")
-        self._outputs_path = self.outputs_folder / "output.jsonl"
+        self._outputs_path = self.outputs_folder / "output.jsonl"  # dumped by executor
         self._node_infos_folder = self._prepare_folder(self.path / "node_artifacts")
         self._run_infos_folder = self._prepare_folder(self.path / "flow_artifacts")
         self._data_path = Path(run.data) if run.data is not None else None
@@ -215,7 +218,7 @@ class LocalStorageOperations(AbstractRunStorage):
 
     def _dump_meta_file(self) -> None:
         with open(self._meta_path, mode="w", encoding=DEFAULT_ENCODING) as f:
-            json.dump({"batch_size": LOCAL_STORAGE_BATCH_SIZE}, f)
+            json.dump({"batch_size": LOCAL_STORAGE_BATCH_SIZE}, f, ensure_ascii=False)
 
     def dump_snapshot(self, flow: Flow) -> None:
         """Dump flow directory to snapshot folder, input file will be dumped after the run."""
@@ -246,28 +249,10 @@ class LocalStorageOperations(AbstractRunStorage):
             flow_dag = yaml.safe_load(f)
         return flow_dag["inputs"], flow_dag["outputs"]
 
-    def dump_inputs(self, line_results: List[LineResult]) -> None:
-        inputs = []
-        for line_result in line_results:
-            try:
-                inputs.append(line_result.run_info.inputs)
-            except Exception:
-                # ignore when single line doesn't have inputs
-                pass
-        df = pd.DataFrame(inputs)
-        with open(self._inputs_path, mode="w", encoding=DEFAULT_ENCODING) as f:
-            # policy: http://policheck.azurewebsites.net/Pages/TermInfo.aspx?LCID=9&TermID=203588
-            df.to_json(f, "records", lines=True)
-
     def load_inputs(self) -> RunInputs:
-        with open(self._inputs_path, mode="r", encoding=DEFAULT_ENCODING) as f:
+        with open(self._sdk_inputs_path, mode="r", encoding=DEFAULT_ENCODING) as f:
             df = pd.read_json(f, orient="records", lines=True)
             return df.to_dict("list")
-
-    def dump_outputs(self, outputs: RunOutputs) -> None:
-        df = pd.DataFrame(outputs)
-        with open(self._outputs_path, mode="w", encoding=DEFAULT_ENCODING) as f:
-            df.to_json(f, "records", lines=True)
 
     def load_outputs(self) -> RunOutputs:
         # for legacy run, simply read the output file and return as list of dict
@@ -287,10 +272,17 @@ class LocalStorageOperations(AbstractRunStorage):
                 df = df.set_index(LINE_NUMBER)
             return df.to_dict("list")
 
+    def dump_inputs_and_outputs(self) -> None:
+        inputs, outputs = self._collect_io_from_debug_info()
+        with open(self._sdk_inputs_path, mode="w", encoding=DEFAULT_ENCODING) as f:
+            inputs.to_json(f, orient="records", lines=True, force_ascii=False)
+        with open(self._sdk_output_path, mode="w", encoding=DEFAULT_ENCODING) as f:
+            outputs.to_json(f, orient="records", lines=True, force_ascii=False)
+
     def dump_metrics(self, metrics: Optional[RunMetrics]) -> None:
         metrics = metrics or dict()
         with open(self._metrics_path, mode="w", encoding=DEFAULT_ENCODING) as f:
-            json.dump(metrics, f)
+            json.dump(metrics, f, ensure_ascii=False)
 
     def dump_exception(self, exception: Exception, bulk_results: BulkResult) -> None:
         """Dump exception to local storage.
@@ -342,7 +334,9 @@ class LocalStorageOperations(AbstractRunStorage):
                 line_errors={"errors": errors},
             )
         with open(self._exception_path, mode="w", encoding=DEFAULT_ENCODING) as f:
-            json.dump(PromptflowExceptionPresenter.create(exception).to_dict(include_debug_info=True), f)
+            json.dump(
+                PromptflowExceptionPresenter.create(exception).to_dict(include_debug_info=True), f, ensure_ascii=False
+            )
 
     def load_exception(self) -> Dict:
         try:
@@ -411,9 +405,8 @@ class LocalStorageOperations(AbstractRunStorage):
         """Persist metrics from return of executor."""
         if result is None:
             return
-        # The executor will persist outputs to output directory, so only dump metrics here for the time being.
+        self.dump_inputs_and_outputs()
         self.dump_metrics(result.metrics)
-        self.dump_inputs(result.line_results)
 
     def _persist_run_multimedia(self, run_info: Union[FlowRunInfo, NodeRunInfo], folder_path: Path):
         if run_info.inputs:
@@ -448,3 +441,35 @@ class LocalStorageOperations(AbstractRunStorage):
         res = pd.concat([df, df_to_append], ignore_index=True)
         res = res.sort_values(by=LINE_NUMBER, ascending=True)
         return res
+
+    def load_inputs_and_outputs(self) -> Tuple[pd.DataFrame, pd.DataFrame]:
+        if not self._sdk_inputs_path.is_file() or not self._sdk_output_path.is_file():
+            inputs, outputs = self._collect_io_from_debug_info()
+        else:
+            with open(self._sdk_inputs_path, mode="r", encoding=DEFAULT_ENCODING) as f:
+                inputs = pd.read_json(f, orient="records", lines=True)
+            with open(self._sdk_output_path, mode="r", encoding=DEFAULT_ENCODING) as f:
+                outputs = pd.read_json(f, orient="records", lines=True)
+                # if all line runs are failed, no need to fill
+                if len(outputs) > 0:
+                    outputs = self._outputs_padding(outputs, len(inputs))
+                    outputs.fillna(value="(Failed)", inplace=True)  # replace nan with explicit prompt
+                    outputs = outputs.set_index(LINE_NUMBER)
+        return inputs, outputs
+
+    def _collect_io_from_debug_info(self) -> Tuple[pd.DataFrame, pd.DataFrame]:
+        inputs, outputs = [], []
+        for line_run_record_file in sorted(self._run_infos_folder.iterdir()):
+            if line_run_record_file.suffix.lower() != ".jsonl":
+                continue
+            with open(line_run_record_file, mode="r", encoding=DEFAULT_ENCODING) as f:
+                data = json.load(f)
+                line_number: int = data[LINE_NUMBER]
+                line_run_info: dict = data["run_info"]
+                current_inputs = line_run_info.get("inputs")
+                current_outputs = line_run_info.get("output")
+                inputs.append(copy.deepcopy(current_inputs))
+                if current_outputs is not None:
+                    current_outputs[LINE_NUMBER] = line_number
+                    outputs.append(copy.deepcopy(current_outputs))
+        return pd.DataFrame(inputs), pd.DataFrame(outputs)